+++ History.md
@@ -1,3 +1,5 @@
+* HeathS: WIXFEAT:4278 - Support redirectable package cache via policy.
+
 * RobMen: WIXBUG:4243 - use hashes to verify bundled packages by default.
 
 * SeanHall: WIXFEAT:3736 - Add WixBundleExecutePackageCacheFolder variable.